Separated global styles into component files
