+++ src/components/TextBlock.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import {
     Clipboard,
+    StyleSheet,
     Text,
     ToastAndroid,
     TouchableOpacity,
     Vibration
 } from 'react-native';
-
-import styles from './styles';
 
 const TextBlock = ({ text }) => {
     if (text) {
@@ -23,13 +22,27 @@ const TextBlock = ({ text }) => {
                     Clipboard.setString(text);
                 }}
             >
-                <Text style={{ flex: 1, fontFamily: 'Roboto_Mono' }}>
-                    {text}
-                </Text>
+                <Text style={styles.textBlockText}>{text}</Text>
             </TouchableOpacity>
         );
     }
     return null;
 };
+
+const styles = StyleSheet.create({
+    textBlock: {
+        backgroundColor: '#f2f2f2',
+        marginTop: 5,
+        marginBottom: 5,
+        paddingLeft: 10,
+        paddingRight: 10,
+        paddingTop: 5,
+        paddingBottom: 5
+    },
+    textBlockText: {
+        flex: 1,
+        fontFamily: 'Roboto_Mono'
+    }
+});
 
 export default TextBlock;
