+++ src/components/containers/BasicDataContainer.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import { Image, Text, TouchableOpacity, View } from 'react-native';
+import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
 import FlightProgressBar from '../FlightProgressBar';
-import styles from '../styles';
 
 const BasicDataContainer = props => (
     <TouchableOpacity
@@ -51,5 +50,48 @@ const BasicDataContainer = props => (
         </View>
     </TouchableOpacity>
 );
+
+const styles = StyleSheet.create({
+    callsignText: {
+        textAlign: 'right',
+        marginRight: 30,
+        fontSize: 13
+    },
+    fromToIcon: {
+        width: 45,
+        height: 45,
+        transform: [
+            {
+                rotate: '90deg'
+            }
+        ]
+    },
+    icaoText: {
+        flex: 1,
+        fontFamily: 'Roboto_Condensed_Regular',
+        fontSize: 50
+    },
+    infoContainerBasic: {
+        flex: 8
+    },
+    infoRow: {
+        flex: 1,
+        flexDirection: 'row',
+        justifyContent: 'center',
+        alignItems: 'center'
+    },
+    nameText: {
+        textAlign: 'left',
+        fontSize: 13
+    },
+    pilotInfoView: {
+        flexDirection: 'row',
+        height: 20
+    },
+    text: {
+        fontFamily: 'Roboto_Regular',
+        color: 'white'
+    }
+});
 
 export default BasicDataContainer;
