+++ src/components/screens/MapScreen.js
@@ -1,16 +1,15 @@
 import React from 'react';
-import { Animated, BackHandler, Text, View } from 'react-native';
+import { Animated, BackHandler, StyleSheet, Text, View } from 'react-native';
 import { ActivityIndicator } from 'react-native-paper';
 import Icon from '@expo/vector-icons/MaterialCommunityIcons';
 
 import MapContainer from '../containers/MapContainer';
 import InfoPanelContainer from '../containers/InfoPanelContainer';
+import { accent as accentColor } from '../../config/colors.json';
 import {
     panelStates,
     panelTransitionDuration
 } from '../../config/constants.json';
-import styles from '../styles';
-import colors from '../../config/colors.json';
 
 export default class MapScreen extends React.PureComponent {
     state = {
@@ -73,7 +72,7 @@ export default class MapScreen extends React.PureComponent {
                     collapsePanel={this.collapsePanel}
                 />
                 <ActivityIndicator
-                    color={colors.accent}
+                    color={accentColor}
                     animating={this.props.screenProps.loading}
                     style={styles.activityIndicator}
                 />
@@ -89,3 +88,17 @@ export default class MapScreen extends React.PureComponent {
         );
     }
 }
+
+const styles = StyleSheet.create({
+    activityIndicator: {
+        position: 'absolute',
+        right: 3,
+        top: 3
+    },
+    clientCountText: {
+        fontFamily: 'Roboto_Regular',
+        position: 'absolute',
+        left: 5,
+        top: 2
+    }
+});
